+++ crypto/src/signature.rs
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2023 TriliTech <contact@trili.tech>
+// SPDX-FileCopyrightText: 2023-2024 Trilitech <contact@trili.tech>
 // Copyright (c) SimpleStaking, Viable Systems and Tezedge Contributors
 //
 // Ported from octez: lib_crypto/signature_v1.ml
@@ -11,10 +11,11 @@
 
 use crate::base58::FromBase58CheckError;
 use crate::hash::{
-    BlsSignature, Ed25519Signature, FromBytesError, HashTrait, P256Signature, Secp256k1Signature,
-    UnknownSignature,
+    BlsSignature, Ed25519Signature, FromBytesError, HashTrait, HashType, P256Signature,
+    Secp256k1Signature, UnknownSignature,
 };
 use serde::{Deserialize, Serialize};
+use thiserror::Error;
 
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize)]
 #[serde(untagged)]
@@ -50,6 +51,16 @@ impl Signature {
             Self::P256(s) => s.to_b58check(),
             Self::Bls(s) => s.to_b58check(),
             Self::Unknown(s) => s.to_b58check(),
+        }
+    }
+
+    pub fn hash_type(&self) -> HashType {
+        match self {
+            Self::Ed25519(_) => HashType::Ed25519Signature,
+            Self::Secp256k1(_) => HashType::Secp256k1Signature,
+            Self::P256(_) => HashType::P256Signature,
+            Self::Bls(_) => HashType::BlsSignature,
+            Self::Unknown(_) => HashType::UnknownSignature,
         }
     }
 }
@@ -122,6 +133,71 @@ impl ::std::fmt::Display for Signature {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         // TODO - this could be done without the need to perform a heap allocation.
         write!(f, "{}", self.to_base58_check())
+    }
+}
+
+macro_rules! from_s_for_sig {
+    ($sig:ident, $name:ident) => {
+        impl From<$sig> for Signature {
+            fn from(s: $sig) -> Self {
+                Self::$name(s)
+            }
+        }
+    };
+}
+from_s_for_sig!(Ed25519Signature, Ed25519);
+from_s_for_sig!(Secp256k1Signature, Secp256k1);
+from_s_for_sig!(P256Signature, P256);
+from_s_for_sig!(BlsSignature, Bls);
+from_s_for_sig!(UnknownSignature, Unknown);
+
+#[derive(Debug, Error)]
+pub enum TryFromSignatureError {
+    #[error("Incorrect signature kind {0:?}.")]
+    InvalidKind(HashType),
+}
+
+macro_rules! from_sig_for_s {
+    ($sig:ident, $name:ident) => {
+        impl TryFrom<Signature> for $sig {
+            type Error = TryFromSignatureError;
+
+            fn try_from(s: Signature) -> Result<Self, Self::Error> {
+                match s {
+                    Signature::$name(s) => Ok(s),
+                    Signature::Unknown(s) => Ok(s.into()),
+                    s => Err(Self::Error::InvalidKind(s.hash_type())),
+                }
+            }
+        }
+    };
+}
+from_sig_for_s!(Ed25519Signature, Ed25519);
+from_sig_for_s!(Secp256k1Signature, Secp256k1);
+from_sig_for_s!(P256Signature, P256);
+
+impl TryFrom<Signature> for BlsSignature {
+    type Error = TryFromSignatureError;
+
+    fn try_from(s: Signature) -> Result<Self, Self::Error> {
+        match s {
+            Signature::Bls(s) => Ok(s),
+            s => Err(Self::Error::InvalidKind(s.hash_type())),
+        }
+    }
+}
+
+impl TryFrom<Signature> for UnknownSignature {
+    type Error = TryFromSignatureError;
+
+    fn try_from(s: Signature) -> Result<Self, Self::Error> {
+        match s {
+            Signature::Ed25519(s) => Ok(s.into()),
+            Signature::Secp256k1(s) => Ok(s.into()),
+            Signature::P256(s) => Ok(s.into()),
+            Signature::Unknown(s) => Ok(s),
+            s => Err(Self::Error::InvalidKind(s.hash_type())),
+        }
     }
 }